Refactor: SearchBox/index.js에서 eventHandler분리
